% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{DkubeCode}
\alias{DkubeCode}
\title{Dkube Code repo}
\usage{
DkubeCode(user = NULL, name = NULL)
}
\arguments{
\item{user}{username}

\item{name}{repo/resource name}

\item{description}{additional details for the repo}
}
\description{
Dkube Code repo
}
