% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{DkubeDataset}
\alias{DkubeDataset}
\title{Dkube Dataset repo}
\usage{
DkubeDataset(user = NULL, name = NULL)
}
\arguments{
\item{user}{username}

\item{name}{repo/resource name}

\item{description}{additional details for the repo}
}
\description{
Dkube Dataset repo
}
