% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{DkubeFeatureSet}
\alias{DkubeFeatureSet}
\title{Dkube Featureset}
\usage{
DkubeFeatureSet(user = NULL, name = NULL)
}
\arguments{
\item{user}{username}

\item{name}{repo/resource name}

\item{description}{additional details for the repo}
}
\description{
Dkube Featureset
}
