% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{DkubeProject}
\alias{DkubeProject}
\title{Dkube Project}
\usage{
DkubeProject(user = NULL, name = NULL)
}
\arguments{
\item{user}{username}

\item{name}{repo/resource name}

\item{description}{additional details for the repo}
}
\description{
Dkube Project
}
