% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{DkubeServing}
\alias{DkubeServing}
\title{Dkube Serving job}
\usage{
DkubeServing(user = NULL, name = NULL, description = NULL)
}
\arguments{
\item{user}{username}

\item{name}{job name}

\item{description}{additional details for the run}
}
\description{
Dkube Serving job
}
