% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{DkubeTraining}
\alias{DkubeTraining}
\title{Dkube Training job}
\usage{
DkubeTraining(user = NULL, name = NULL, description = NULL)
}
\arguments{
\item{user}{username}

\item{name}{job name}

\item{description}{additional details for the run}
}
\description{
Dkube Training job
}
